[26684] Redirect to WP from search only with #ID match on enter
Currently, the backend redirects to work packages whenever an ID is
exactly matched with or without the `#` prefix. This means users cannot
explicitly trigger a search through the "global" or "in this project"
buttons.

If we remember the match in the frontend, we can redirect it _only_ when
pressing enter. I also restricted the search to match with the `#`
prefix only

https://community.openproject.com/wp/26684